backend/requirements.txt
backports.tarfile==1.2.0
bottleneck==1.4.2
brotli==1.0.9
conda-lock==2.5.6
fastapi==0.112.2
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.collections==5.1.0
joblib==1.4.2
jupyter==1.0.0
neo4j==5.27.dev0
numexpr==2.10.1
pip-chill==1.0.3
pyopenssl==24.2.1
pyqt5==5.15.10
pysocks==1.7.1
pytest==7.4.4
python-dotenv==0.21.0
ruamel.yaml==0.18.6
scipy==1.15.1
seaborn==0.13.2
sip==6.7.12
streamlit==1.41.1
tomli==2.0.1
tqdm==4.66.5
unicodedata2==15.1.0
uvicorn==0.32.1
watchdog==4.0.1
